Introduce t.Parallel() for some tests
Also:

* removed build tags
* properly renamed test files to blackbox and whitebox tests
