+++ models/workitemtype_whitebox_test.go
@@ -9,6 +9,7 @@ import (
 )
 
 func TestConvertFieldTypes(t *testing.T) {
+	t.Parallel()
 	resource.Require(t, resource.UnitTest)
 	types := []FieldType{
 		SimpleType{Kind: KindInteger},
